// Alamin_S361035/week_9/CareerCenterAppUI/RegistrationApp.UI/src/app/models/students.model.ts
export interface Student{
    StudentRegistrationId: number
    StudentName: string;
    PreferredName: string;
    StudentID: string;
    Course: string;
    CurrentSemester: number;
    Email: string;
    PhoneNumber: number;
    ReasonsToJoin: string;
    ExpectedOutcome: string;
    TimeAvailability: number;
    PhotographyConsent: number;
}